+++ src/main/java/server/life/MonsterListener.java
@@ -5,6 +5,6 @@ import client.Character;
 public interface MonsterListener {
 
     void monsterKilled(int aniTime);
-    void monsterDamaged(Character from, int trueDmg);
-    void monsterHealed(int trueHeal);
+    void monsterDamaged(Character from, long trueDmg);
+    void monsterHealed(long trueHeal);
 }
